Task 2/requirements.txt
numpy==1.26.0
openai==1.107.0
python-dotenv==1.1.1
streamlit==1.48.1
requests>=2.32.3
gensim==4.3.3